Rework driver object
Require three "methods" for a driver: `open`, `close` and `poll`. The
`open` method is expected to open a device for polling and return `true`
if succesful. The `close` method is expected to close/free any associated
resources used for polling and to make sure the device can be cleanly
opened again for future polling.