Add setting of "capabilities" of devices on scanning
For each device found, determine if it can be used to emulate a
joystick, a mouse, a koala pad or a paddle. Currently the linux-specific
code calls `joy_device_set_capabilities()`, perhaps a arch-independent
wrapper should be created that calls the arch-specific device scanning
code and then iterates that list to set the capabilities.